code clean up

+++ src/Controller/AppelationController.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\Controller;
 
 use App\Entity\Appelation;
@@ -7,6 +9,7 @@ use App\Form\AppelationForm;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\Form\Extension\Core\Type\SubmitType;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Security\Http\Attribute\IsGranted;
 
@@ -17,9 +20,9 @@ class AppelationController extends AbstractController
      * affiche le tableau de bord de gestion des appelations.
      */
     #[Route('/appelation', name: 'appelation.index')]
-    public function indexAction(Request $request, EntityManagerInterface $entityManager): \Symfony\Component\HttpFoundation\Response
+    public function indexAction(Request $request, EntityManagerInterface $entityManager): Response
     {
-        $appelations = $entityManager->getRepository('\\'.Appelation::class)->findAll();
+        $appelations = $entityManager->getRepository('\\' . Appelation::class)->findAll();
         $appelations = $this->sortAppelation($appelations);
 
         return $this->render('appelation/index.twig', ['appelations' => $appelations]);
@@ -29,8 +32,11 @@ class AppelationController extends AbstractController
      * Detail d'une appelation.
      */
     #[Route('/appelation/{appelation}/detail', name: 'appelation.detail')]
-    public function detailAction(Request $request, EntityManagerInterface $entityManager, Appelation $appelation): \Symfony\Component\HttpFoundation\Response
-    {
+    public function detailAction(
+        Request $request,
+        EntityManagerInterface $entityManager,
+        Appelation $appelation,
+    ): Response {
         return $this->render('appelation/detail.twig', ['appelation' => $appelation]);
     }
 
@@ -38,11 +44,11 @@ class AppelationController extends AbstractController
      * Ajoute une appelation.
      */
     #[Route('/appelation/add', name: 'appelation.add')]
-    public function addAction(Request $request, EntityManagerInterface $entityManager)
+    public function addAction(Request $request, EntityManagerInterface $entityManager): Response
     {
-        $form = $this->createForm(AppelationForm::class, new Appelation())
-            ->add('save', SubmitType::class, ['label' => 'Sauvegarder'])
-            ->add('save_continue', SubmitType::class, ['label' => 'Sauvegarder & continuer']);
+        $form = $this->createForm(AppelationForm::class, new Appelation())->add('save', SubmitType::class, [
+            'label' => 'Sauvegarder',
+        ])->add('save_continue', SubmitType::class, ['label' => 'Sauvegarder & continuer']);
 
         $form->handleRequest($request);
 
@@ -53,9 +59,9 @@ class AppelationController extends AbstractController
 
             $this->addFlash('success', 'L\'appelation a été ajoutée.');
 
-            if ($form->get('save')->isClicked()) {
+            if ($form->get('save') instanceof \Symfony\Component\Form\ClickableInterface && $form->get('save')->isClicked()) {
                 return $this->redirectToRoute('appelation.index', [], 303);
-            } elseif ($form->get('save_continue')->isClicked()) {
+            } elseif ($form->get('save_continue') instanceof \Symfony\Component\Form\ClickableInterface && $form->get('save_continue')->isClicked()) {
                 return $this->redirectToRoute('appelation.add', [], 303);
             }
         }
@@ -69,25 +75,28 @@ class AppelationController extends AbstractController
      * Modifie une appelation.
      */
     #[Route('/appelation/{appelation}/update', name: 'appelation.update')]
-    public function updateAction(Request $request, EntityManagerInterface $entityManager, Appelation $appelation)
-    {
-        $form = $this->createForm(AppelationForm::class, $appelation)
-            ->add('update', SubmitType::class, ['label' => 'Sauvegarder'])
-            ->add('delete', SubmitType::class, ['label' => 'Supprimer']);
+    public function updateAction(
+        Request $request,
+        EntityManagerInterface $entityManager,
+        Appelation $appelation,
+    ): Response {
+        $form = $this->createForm(AppelationForm::class, $appelation)->add('update', SubmitType::class, [
+            'label' => 'Sauvegarder',
+        ])->add('delete', SubmitType::class, ['label' => 'Supprimer']);
 
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $appelation = $form->getData();
 
-            if ($form->get('update')->isClicked()) {
+            if ($form->get('update') instanceof \Symfony\Component\Form\ClickableInterface && $form->get('update')->isClicked()) {
                 $entityManager->persist($appelation);
                 $entityManager->flush();
                 $this->addFlash('success', 'L\'appelation a été mise à jour.');
 
                 return $this->redirectToRoute('appelation.detail', ['appelation' => $appelation->getId()], 303);
             }
-            if ($form->get('delete')->isClicked()) {
+            if ($form->get('delete') instanceof \Symfony\Component\Form\ClickableInterface && $form->get('delete')->isClicked()) {
                 $entityManager->remove($appelation);
                 $entityManager->flush();
                 $this->addFlash('success', 'L\'appelation a été supprimée.');
@@ -105,7 +114,9 @@ class AppelationController extends AbstractController
     /**
      * Classement des appelations par groupe.
      *
-     * @return array $appelations
+     * @param array<int, Appelation> $appelations
+     *
+     * @return array<int, Appelation>
      */
     public function sortAppelation(array $appelations): array
     {
@@ -115,17 +126,16 @@ class AppelationController extends AbstractController
         // recherche des racines (appelations n'ayant pas de parent
         // dans la liste des appelations fournis)
         foreach ($appelations as $appelation) {
-            if (!in_array($appelation->getAppelation(), $appelations, true)) {
-                $root[] = $appelation;
+            if (\in_array($appelation->getAppelation(), $appelations, true)) {
+                continue;
             }
+
+            $root[] = $appelation;
         }
 
         foreach ($root as $appelation) {
-            if (count($appelation->getAppelations()) > 0) {
-                $childs = array_merge(
-                    [$appelation],
-                    $this->sortAppelation($appelation->getAppelations()->toArray())
-                );
+            if (\count($appelation->getAppelations()) > 0) {
+                $childs = array_merge([$appelation], $this->sortAppelation($appelation->getAppelations()->toArray()));
 
                 $result = [...$result, ...$childs];
             } else {
