+++ src/Controller/LangueController.php
@@ -1,30 +1,28 @@
 <?php
 
+declare(strict_types=1);
 
 namespace App\Controller;
 
 use App\Entity\GroupeLangue;
 use App\Entity\Langue;
-use App\Form\LangueForm;
 use App\Form\Groupe\GroupeLangueForm;
-use App\Repository\LangueRepository;
-use App\Service\PagerService;
+use App\Form\LangueForm;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\Form\Extension\Core\Type\SubmitType;
-use Symfony\Component\Form\Form;
-use Symfony\Component\HttpFoundation\File\Exception\FileException;
+use Symfony\Component\Form\FormInterface;
 use Symfony\Component\HttpFoundation\BinaryFileResponse;
+use Symfony\Component\HttpFoundation\File\Exception\FileException;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\HttpFoundation\ResponseHeaderBag;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Security\Http\Attribute\IsGranted;
 
-#[isGranted('ROLE_SCENARISTE')]
+#[IsGranted('ROLE_SCENARISTE')]
 class LangueController extends AbstractController
 {
-    final public const DOC_PATH = __DIR__.'/../../private/doc/';
+    final public const DOC_PATH = __DIR__ . '/../../private/doc/';
 
     /**
      * affiche la liste des langues.
@@ -51,13 +49,13 @@ class LangueController extends AbstractController
      * Ajoute une langue.
      */
     #[Route('/langue/add', name: 'langue.add')]
-    public function addAction(Request $request,  EntityManagerInterface $entityManager): Response|RedirectResponse
+    public function addAction(Request $request, EntityManagerInterface $entityManager): Response|RedirectResponse
     {
         $langue = new Langue();
 
-        $form = $this->createForm(LangueForm::class, $langue)
-            ->add('save', SubmitType::class, ['label' => 'Sauvegarder'])
-            ->add('save_continue', SubmitType::class, ['label' => 'Sauvegarder & continuer']);
+        $form = $this->createForm(LangueForm::class, $langue)->add('save', SubmitType::class, [
+            'label' => 'Sauvegarder',
+        ])->add('save_continue', SubmitType::class, ['label' => 'Sauvegarder & continuer']);
 
         $form->handleRequest($request);
 
@@ -68,13 +66,13 @@ class LangueController extends AbstractController
                 $entityManager->persist($langue);
                 $entityManager->flush();
 
-               $this->addFlash('success', 'La langue a été ajoutée.');
+                $this->addFlash('success', 'La langue a été ajoutée.');
 
                 // l'utilisateur est redirigé soit vers la liste des langues, soit vers de nouveau
                 // vers le formulaire d'ajout d'une langue
-                if ($form->get('save')->isClicked()) {
+                if ($form->get('save') instanceof \Symfony\Component\Form\ClickableInterface && $form->get('save')->isClicked()) {
                     return $this->redirectToRoute('langue.index', [], 303);
-                } elseif ($form->get('save_continue')->isClicked()) {
+                } elseif ($form->get('save_continue') instanceof \Symfony\Component\Form\ClickableInterface && $form->get('save_continue')->isClicked()) {
                     return $this->redirectToRoute('langue.add', [], 303);
                 }
             }
@@ -92,18 +90,31 @@ class LangueController extends AbstractController
      * redirigé vers la liste des langues.
      */
     #[Route('/langue/{langue}/update', name: 'langue.update')]
-    public function updateAction(Request $request,  EntityManagerInterface $entityManager, Langue $langue): Response|RedirectResponse
-    {
+    public function updateAction(
+        Request $request,
+        EntityManagerInterface $entityManager,
+        Langue $langue,
+    ): Response|RedirectResponse {
         $hasDocumentUrl = !empty($langue->getDocumentUrl());
-        $canBeDeleted = $langue->getPersonnageLangues()->isEmpty()
-            && $langue->getTerritoires()->isEmpty()
-            && $langue->getDocuments()->isEmpty();
+        $canBeDeleted = $langue->getPersonnageLangues()->isEmpty() && $langue->getTerritoires()->isEmpty() && $langue->getDocuments()->isEmpty();
 
-        $deleteTooltip = $canBeDeleted ? '' : 'Cette langue est référencée par '.$langue->getPersonnageLangues()->count().' personnages, '.$langue->getTerritoires()->count().' territoires et '.$langue->getDocuments()->count().' documents et ne peut pas être supprimée';
+        $deleteTooltip = $canBeDeleted
+            ? ''
+            : 'Cette langue est référencée par '
+            . $langue->getPersonnageLangues()->count()
+            . ' personnages, '
+            . $langue->getTerritoires()->count()
+            . ' territoires et '
+            . $langue->getDocuments()->count()
+            . ' documents et ne peut pas être supprimée';
 
-        $formBuilder = $this->createForm(LangueForm::class, $langue, ['hasDocumentUrl' => $hasDocumentUrl])
-            ->add('update', SubmitType::class, ['label' => 'Sauvegarder'])
-            ->add('delete', SubmitType::class, ['label' => 'Supprimer', 'disabled' => !$canBeDeleted, 'attr' => ['title' => $deleteTooltip]]);
+        $formBuilder = $this->createForm(LangueForm::class, $langue, [
+            'hasDocumentUrl' => $hasDocumentUrl,
+        ])->add('update', SubmitType::class, ['label' => 'Sauvegarder'])->add('delete', SubmitType::class, [
+            'label' => 'Supprimer',
+            'disabled' => !$canBeDeleted,
+            'attr' => ['title' => $deleteTooltip],
+        ]);
 
         $form = $formBuilder;
 
@@ -112,7 +123,7 @@ class LangueController extends AbstractController
         if ($form->isSubmitted() && $form->isValid()) {
             $langue = $form->getData();
 
-            if ($form->get('update')->isClicked()) {
+            if ($form->get('update') instanceof \Symfony\Component\Form\ClickableInterface && $form->get('update')->isClicked()) {
                 if (self::handleDocument($request, $entityManager, $form, $langue)) {
                     $entityManager->persist($langue);
                     $entityManager->flush();
@@ -120,12 +131,12 @@ class LangueController extends AbstractController
 
                     return $this->redirectToRoute('langue.detail', ['langue' => $langue->getId()], 303);
                 }
-            } elseif ($form->get('delete')->isClicked()) {
+            } elseif ($form->get('delete') instanceof \Symfony\Component\Form\ClickableInterface && $form->get('delete')->isClicked()) {
                 $entityManager->remove($langue);
                 $entityManager->flush();
                 // delete language document if it exists
                 self::tryDeleteDocument($langue);
-               $this->addFlash('success', 'La langue a été supprimée.');
+                $this->addFlash('success', 'La langue a été supprimée.');
 
                 return $this->redirectToRoute('langue.index', [], 303);
             }
@@ -140,22 +151,26 @@ class LangueController extends AbstractController
     /**
      * Gère le document uploadé et renvoie true si il est valide, false sinon.
      */
-    private function handleDocument(Request $request,  EntityManagerInterface $entityManager, Form $form, Langue $langue): bool
-    {
+    private function handleDocument(
+        Request $request,
+        EntityManagerInterface $entityManager,
+        FormInterface $form,
+        Langue $langue,
+    ): bool {
         $files = $request->files->get($form->getName());
         $documentFile = $files['document'];
         // Si un document est fourni, l'enregistrer
         if (null != $documentFile) {
             $filename = $documentFile->getClientOriginalName();
-            $extension = pathinfo((string) $filename, PATHINFO_EXTENSION);
+            $extension = pathinfo((string) $filename, \PATHINFO_EXTENSION);
 
             if (!$extension || 'pdf' != $extension) {
-               $this->addFlash('error', 'Désolé, votre document n\'est pas valide. Vérifiez le format de votre document ('.$extension.'), seuls les .pdf sont acceptés.');
+                $this->addFlash('error', 'Désolé, votre document n\'est pas valide. Vérifiez le format de votre document (' . $extension . '), seuls les .pdf sont acceptés.');
 
                 return false;
             }
 
-            $documentFilename = hash('md5', $langue->getLabel().$filename.time()).'.'.$extension;
+            $documentFilename = hash('md5', $langue->getLabel() . $filename . time()) . '.' . $extension;
 
             $documentFile->move(self::DOC_PATH, $documentFilename);
 
@@ -177,7 +192,7 @@ class LangueController extends AbstractController
     {
         try {
             if (!empty($langue->getDocumentUrl())) {
-                $docFilePath = self::DOC_PATH.$langue->getDocumentUrl();
+                $docFilePath = self::DOC_PATH . $langue->getDocumentUrl();
                 unlink($docFilePath);
             }
         } catch (FileException) {
@@ -189,8 +204,11 @@ class LangueController extends AbstractController
      * Detail d'un groupe de langue.
      */
     #[Route('/langue/groupe/{groupeLangue}/detail', name: 'langue.detailGroup')]
-    public function detailGroupAction(Request $request,  EntityManagerInterface $entityManager, GroupeLangue $groupeLangue): Response
-    {
+    public function detailGroupAction(
+        Request $request,
+        EntityManagerInterface $entityManager,
+        GroupeLangue $groupeLangue,
+    ): Response {
         return $this->render('langue/detailGroup.twig', ['groupeLangue' => $groupeLangue]);
     }
 
@@ -198,13 +216,13 @@ class LangueController extends AbstractController
      * Ajoute un groupe de langue.
      */
     #[Route('/langue/groupe/add', name: 'langue.addGroup')]
-    public function addGroupAction(Request $request,  EntityManagerInterface $entityManager): Response|RedirectResponse
+    public function addGroupAction(Request $request, EntityManagerInterface $entityManager): Response|RedirectResponse
     {
         $groupeLangue = new GroupeLangue();
 
-        $form = $this->createForm(GroupeLangueForm::class, $groupeLangue)
-            ->add('save', SubmitType::class, ['label' => 'Sauvegarder'])
-            ->add('save_continue', SubmitType::class, ['label' => 'Sauvegarder & continuer']);
+        $form = $this->createForm(GroupeLangueForm::class, $groupeLangue)->add('save', SubmitType::class, [
+            'label' => 'Sauvegarder',
+        ])->add('save_continue', SubmitType::class, ['label' => 'Sauvegarder & continuer']);
 
         $form->handleRequest($request);
 
@@ -214,13 +232,13 @@ class LangueController extends AbstractController
             $entityManager->persist($groupeLangue);
             $entityManager->flush();
 
-           $this->addFlash('success', 'Le groupe de langue a été ajouté.');
+            $this->addFlash('success', 'Le groupe de langue a été ajouté.');
 
             // l'utilisateur est redirigé soit vers la liste des langues, soit vers de nouveau
             // vers le formulaire d'ajout d'une langue
-            if ($form->get('save')->isClicked()) {
+            if ($form->get('save') instanceof \Symfony\Component\Form\ClickableInterface && $form->get('save')->isClicked()) {
                 return $this->redirectToRoute('langue.index', [], 303);
-            } elseif ($form->get('save_continue')->isClicked()) {
+            } elseif ($form->get('save_continue') instanceof \Symfony\Component\Form\ClickableInterface && $form->get('save_continue')->isClicked()) {
                 return $this->redirectToRoute('langue.addGroup', [], 303);
             }
         }
@@ -236,14 +254,23 @@ class LangueController extends AbstractController
      * redirigé vers la liste des langues.
      */
     #[Route('/langue/groupe/{groupeLangue}/update', name: 'langue.updateGroup')]
-    public function updateGroupAction(Request $request,  EntityManagerInterface $entityManager, GroupeLangue $groupeLangue): Response|RedirectResponse
-    {
+    public function updateGroupAction(
+        Request $request,
+        EntityManagerInterface $entityManager,
+        GroupeLangue $groupeLangue,
+    ): Response|RedirectResponse {
         $canBeDeleted = $groupeLangue->getLangues()->isEmpty();
-        $deleteTooltip = $canBeDeleted ? '' : 'Ce groupe est référencé par '.$groupeLangue->getLangues()->count().' langues et ne peut pas être supprimé';
+        $deleteTooltip = $canBeDeleted
+            ? ''
+            : 'Ce groupe est référencé par ' . $groupeLangue->getLangues()->count() . ' langues et ne peut pas être supprimé';
 
-        $formBuilder = $this->createForm(GroupeLangueForm::class, $groupeLangue)
-            ->add('update', SubmitType::class, ['label' => 'Sauvegarder'])
-            ->add('delete', SubmitType::class, ['label' => 'Supprimer', 'disabled' => !$canBeDeleted, 'attr' => ['title' => $deleteTooltip]]);
+        $formBuilder = $this->createForm(GroupeLangueForm::class, $groupeLangue)->add('update', SubmitType::class, [
+            'label' => 'Sauvegarder',
+        ])->add('delete', SubmitType::class, [
+            'label' => 'Supprimer',
+            'disabled' => !$canBeDeleted,
+            'attr' => ['title' => $deleteTooltip],
+        ]);
 
         $form = $formBuilder;
 
@@ -252,16 +279,16 @@ class LangueController extends AbstractController
         if ($form->isSubmitted() && $form->isValid()) {
             $groupeLangue = $form->getData();
 
-            if ($form->get('update')->isClicked()) {
+            if ($form->get('update') instanceof \Symfony\Component\Form\ClickableInterface && $form->get('update')->isClicked()) {
                 $entityManager->persist($groupeLangue);
                 $entityManager->flush();
-               $this->addFlash('success', 'Le groupe de langue a été mis à jour.');
+                $this->addFlash('success', 'Le groupe de langue a été mis à jour.');
 
                 return $this->redirectToRoute('langue.detailGroup', ['groupeLangue' => $groupeLangue->getId()], 303);
-            } elseif ($form->get('delete')->isClicked()) {
+            } elseif ($form->get('delete') instanceof \Symfony\Component\Form\ClickableInterface && $form->get('delete')->isClicked()) {
                 $entityManager->remove($groupeLangue);
                 $entityManager->flush();
-               $this->addFlash('success', 'Le groupe de langue a été supprimé.');
+                $this->addFlash('success', 'Le groupe de langue a été supprimé.');
 
                 return $this->redirectToRoute('langue.index', [], 303);
             }
@@ -280,13 +307,13 @@ class LangueController extends AbstractController
     public function adminDocumentAction(Request $request, Langue $langue): BinaryFileResponse
     {
         $document = $langue->getDocumentUrl();
-        $file = self::DOC_PATH.$document;
+        $file = self::DOC_PATH . $document;
 
         $response = new BinaryFileResponse($file);
         $response->headers->set('Content-Control', 'private');
         $response->headers->set('Content-Type', 'text/pdf');
-        $response->headers->set('Content-length', filesize($file));
-        $response->headers->set('Content-Disposition', 'attachment; filename="'.$langue->getPrintLabel().'.pdf"');
+        $response->headers->set('Content-length', (string) filesize($file));
+        $response->headers->set('Content-Disposition', 'attachment; filename="' . $langue->getPrintLabel() . '.pdf"');
 
         return $response;
     }
